Move cells demo into its own page
