+++ extension/source/QuillPage/QuillPage.tsx
@@ -3,8 +3,6 @@ import { HashRouter, Route, Routes } from 'react-router-dom';
 
 import OnboardingPage from './Onboarding/OnboardingPage';
 import { WalletPage } from './Wallet/WalletPage';
-import { CellsDemoPage } from './CellsDemo/CellsDemoPage';
-import { CellsDemoPage2 } from './CellsDemo/CellsDemoPage2';
 import { QuillProvider } from './QuillContext';
 import Theme from './Theme';
 
@@ -16,8 +14,6 @@ const QuillPage: FunctionComponent = () => {
           <Routes>
             <Route path="/onboarding" element={<OnboardingPage />} />
             <Route path="/wallet/*" element={<WalletPage />} />
-            <Route path="/cells-demo" element={<CellsDemoPage />} />
-            <Route path="/cells-demo2" element={<CellsDemoPage2 />} />
           </Routes>
         </HashRouter>
       </Theme>
