+++ extension/source/cells/CellsDemo/CellsDemoPage2.tsx
@@ -1,6 +1,8 @@
 import { FunctionComponent } from 'react';
 import DemoTable from './DemoTable';
 
+// Accessible at chrome-extension://<insert extension id>/cellsDemo.html#/demo2
+
 export const CellsDemoPage2: FunctionComponent = () => {
   return (
     <DemoTable>
