+++ extension/source/cells/CellsDemo/CheckBox.tsx
@@ -1,6 +1,6 @@
 import { FunctionComponent } from 'react';
-import ICell from '../../cells/ICell';
-import useCell from '../../cells/useCell';
+import ICell from '../ICell';
+import useCell from '../useCell';
 
 const CheckBox: FunctionComponent<{ cell: ICell<boolean> }> = ({ cell }) => {
   const value = useCell(cell);
