+++ extension/source/cells/CellsDemo/Display.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from 'react';
 
-import { IReadableCell } from '../../cells/ICell';
-import useCell from '../../cells/useCell';
+import { IReadableCell } from '../ICell';
+import useCell from '../useCell';
 
 export const Display: FunctionComponent<{
   cell: IReadableCell<unknown>;
