+++ extension/source/cells/CellsDemo/Selector.tsx
@@ -1,6 +1,6 @@
 import { FunctionComponent } from 'react';
-import ICell from '../../cells/ICell';
-import useCell from '../../cells/useCell';
+import ICell from '../ICell';
+import useCell from '../useCell';
 
 const Selector: FunctionComponent<{
   options: string[];
