+++ extension/views/cellsDemo.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Cells Demo</title>
+  </head>
+  <body>
+    <div id="cells-demo-root" class="quill"></div>
+  </body>
+</html>
